+++ .commandments/config.php
@@ -33,5 +33,7 @@ return function (Config $config): void {
     );
 
     $config->exclude(
+        'src/web/dist',
+        'src/web/node_modules',
     );
 };